EXPR: move batching logic to the library
The batching logic was earlier being handled by rails `find_in_batches`. The problem with that approach is that it creates multiple Salesforce Bulk Jobs which run in parallel and might conflict with each other causing race conditions on the salesforce records. If this approach works with the Contact type records, we'll use it for other rake tasks as well.